Add testing for variable and function declarations
Add tests to ensure that variable and function declarations parse.
Please note that function pointer declarations with double opening
parentheses are not going to parse. This should be fixed in the future.
There needs to be a `*` after the first parentheses for the type to
parse.
e.g. `int ((*myfunc))()` won't parse
but `int (*((*myfunc)))()` will