Add support for types
Add support for types. These grammar rules should probably be improved
in the future.
As it stands `maybe_typed_name_decl` is used to match all the following:
  - `x` untyped variable declaration
  - `int x` typed variable declaration
  - `int f(int x, y): ...` C functions
  - `MyClass()` C++ constructor declarations

Also keep in mind that an hack is used to match both typed and untyped
variable declarations (similar to `test ['=' test]`), to combat `LL(1)`
grammar limitations. We first match against a type, so in case it's an
ease enter the commit message for your changes. Lines starting